# Use hashids for smaller keys and less memory usage, fix #288