--- a/docs/import.md
+++ b/docs/import.md
@@ -14,7 +14,7 @@ The expected keys are the ones declared in the `FIELDS` attribute of your
 
 - **type** is expected even if not declared as index field
 - **importance** is expected; must be a `float` between 0 and 1
-- **id** is expected
+- **_id** is optional and for internal use only, if not present a [hashid](http://hashids.org/) will be generated; note that shorter keys lead to smaller memory consumption on the Redis side
 - **lat** and **lon** are expected
 - **housenumbers** has a special format: `{number: {lat: yyy, lon: xxx}}` (optionally you can
   add any custom attribute)