--- a/requirements.txt
+++ b/requirements.txt
@@ -1,4 +1,5 @@
 falcon==1.1.0
+hashids==1.2.0
 hiredis==0.2.0
 ngram==3.3.0
 progressist==0.0.6