--- a/tests/test_scripts.py
+++ b/tests/test_scripts.py
@@ -8,8 +8,8 @@ def test_manual_scan(factory):
     street2 = factory(name="rue de la monnaie", city="Condom", importance=0.9)
     results = scripts.manual_scan(keys=['w|monnaie', 'w|rue', 'w|de'],
                                   args=[2])
-    assert results == ['d|{}'.format(street1['id']).encode(),
-                       'd|{}'.format(street2['id']).encode()]
+    assert results == ['d|{}'.format(street1['_id']).encode(),
+                       'd|{}'.format(street2['_id']).encode()]
 
 
 def test_zinter(factory):
@@ -21,10 +21,10 @@ def test_zinter(factory):
     )
     results = scripts.zinter(keys=['w|monnaie', 'w|rue', 'w|de'],
                              args=['tmp', 2])
-    assert results == ['d|{}'.format(docs[2]['id']).encode(),
-                       'd|{}'.format(docs[3]['id']).encode()]
+    assert results == ['d|{}'.format(docs[2]['_id']).encode(),
+                       'd|{}'.format(docs[3]['_id']).encode()]
     results = scripts.zinter(keys=['w|monnaie', 'w|rue', 'w|de'],
                              args=['tmp', 3])
-    assert results == ['d|{}'.format(docs[2]['id']).encode(),
-                       'd|{}'.format(docs[3]['id']).encode(),
-                       'd|{}'.format(docs[0]['id']).encode()]
+    assert results == ['d|{}'.format(docs[2]['_id']).encode(),
+                       'd|{}'.format(docs[3]['_id']).encode(),
+                       'd|{}'.format(docs[0]['_id']).encode()]