Loading `BAN_odbl_76-json`:

Before (master): `Done: 35023 | Elapsed: 0:01:16 | 1m 20s 349ms`

with:

```
> DBINFO
keyspace_misses: 62433
keyspace_hits: 4453
used_memory_human: 297.95M
total_commands_processed: 9859055
total_connections_received: 83
connected_clients: 1
nb keys (db 0): 108358
nb keys (db 1): 35023
```

After: `Done: 35023 | Elapsed: 0:01:08 | 1m 12s 522ms`

with:

```
> DBINFO
keyspace_misses: 62433
keyspace_hits: 4453
used_memory_human: 279.78M
total_commands_processed: 5006893
total_connections_received: 66
connected_clients: 1
nb keys (db 0): 108359
nb keys (db 1): 35023
```

The benchmark is not really relevant from one run to another but we don't have a significative loss. The memory has reduced 🎉 and the difference will be more important with more keys!

I wonder if the `_` (underscore) to prefix `hashid` is pertinent.