@davidbgk 2017-03-15
I don't get the significative loss in coverage given the few lines of code affected. Thoughts?