@davidbgk 2017-03-15
Using `sqlite-store`:

With hashids:

```
> DBINFO
keyspace_misses: 0
keyspace_hits: 0
used_memory_human: 259.64M
total_commands_processed: 2426074
total_connections_received: 4
connected_clients: 1
nb keys (db 0): 108359
```

Without hashids:

```
> DBINFO
keyspace_misses: 0
keyspace_hits: 0
used_memory_human: 276.62M
total_commands_processed: 7243203
total_connections_received: 13
connected_clients: 1
nb keys (db 0): 108358
```