@davidbgk 2017-03-15
To sum up, we have a drop of 6.1% in memory consumption for that dataset, whatever the datastore. It should be even better the more addresses you add.