@yohanboniface 2017-03-15
> Removing the _id prior to save the document doesn't change significantly the memory consumption: used_memory_human: 279.08M (vs. 279.78M above). Do you think it's worth it for cleaner data @yohanboniface ?

Nope, let's keep it then